[refactor] TripRecord 엔티티 '평균 평점' 필드 Double 타입으로 변경 및 그에 따른 연쇄 수정 업데이트
related to: #89

+++ src/main/java/com/haejwo/tripcometrue/domain/store/dto/response/TripRecordStoreResponseDto.java
@@ -12,7 +12,7 @@ public record TripRecordStoreResponseDto(
     LocalDate tripStartDay,
     LocalDate tripEndDay,
     Integer totalDays,
-    Integer averageRating,
+    Double averageRating,
     Integer viewCount
 ) {
